refactor: move assertNever out of test utilities
