+++ shared/helperFunctions.ts
